audio-change-zcr/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
soundfile>=0.10.0
librosa>=0.9.0
scipy>=1.7.0
scikit-learn>=1.0.0
dtw-python>=1.2.0